Reimplementation: docker-bash-into-dead by delegating to docker-execi
docker-bash-into-dead does not support DOCKER-RUN-ARGS (like -v for mounting volumes); we'd have to duplicate the whole parsing from docker-execi.
Instead, add another --revive-dead hook that does a dead container lookup and creates the temp image, and we can reimplement docker-bash-into-dead in terms of docker-execi (and for consistency, also add the intermediate docker-exec-into-dead, too).
CHG: Only the temp image is cleaned up, not the container itself; for that, pass --rm as usual.